move in_range?
