+++ test/dotcom/utils/service_date_time_test.exs
@@ -140,58 +140,6 @@ defmodule Dotcom.Utils.ServiceDateTimeTest do
     end
   end
 
-  describe "in_range?/2" do
-    test "returns false when no actual range is given" do
-      date_time = random_date_time()
-
-      range = {nil, nil}
-
-      refute in_range?(range, date_time)
-    end
-
-    test "returns true when the date_time is the start of the range" do
-      date_time = random_date_time()
-
-      range = {date_time, nil}
-
-      assert in_range?(range, date_time)
-    end
-
-    test "returns true when the date_time is the end of the range" do
-      date_time = random_date_time()
-
-      range = {nil, date_time}
-
-      assert in_range?(range, date_time)
-    end
-
-    property "returns true when the date_time is within the range" do
-      # Setup
-      check all(date_time <- date_time_generator()) do
-        start = Timex.shift(date_time, years: -1)
-        stop = Timex.shift(date_time, years: 1)
-
-        range = {start, stop}
-
-        # Exercise / Verify
-        assert in_range?(range, date_time)
-      end
-    end
-
-    property "returns false when the date_time is not within the range" do
-      # Setup
-      check all(date_time <- date_time_generator()) do
-        start = Timex.shift(date_time, seconds: 1)
-        stop = Timex.shift(start, years: 1)
-
-        range = {start, stop}
-
-        # Exercise / Verify
-        refute in_range?(range, date_time)
-      end
-    end
-  end
-
   describe "service_today?/1" do
     test "returns true when the date_time is in today's service" do
       # Setup
